+++ lib/xcodeproj/project/object/target_dependency.rb
@@ -39,6 +39,17 @@ module Xcodeproj
           return target_proxy.remote_info if target_proxy
         end
 
+        # @return [String] uuid of the target, if the dependency
+        #         is a native target, otherwise the uuid of the
+        #         target in the sub-project if the dependency is
+        #         a target proxy
+        #
+        def native_target_uuid
+          return target.uuid if target
+          return target_proxy.remote_global_id_string if target_proxy
+          raise 'Expected target or target_proxy, from which to fetch a uuid'
+        end
+
         # @note This override is necessary because Xcode allows for circular
         #       target dependencies.
         #